Remove unnecessary methods.
