+++ QRCodeTarot/Card.swift
@@ -8,6 +8,7 @@
 import UIKit
 import Callable
 import ARKit
+import CommonExtensions
 
 // MARK: - Card
 struct Card: Codable, CaseIterable {
