--- QRCodeTarot/Extensions/BinaryInteger.swift
@@ -1,15 +1,0 @@
-//
-//  BinaryInteger.swift
-//  QRCodeTarot
-//
-//  Created by Scott Lydon on 4/9/22.
-//
-
-import Foundation
-
-extension BinaryInteger {
-
-    var isEven: Bool {
-        self % 2 == 0
-    }
-}
