--- QRCodeTarot/Extensions/UICollectionViewFlowLayout.swift
@@ -1,15 +1,0 @@
-//
-//  UICollectionViewFlowLayout.swift
-//  QRCodeTarot
-//
-//  Created by Scott Lydon on 5/4/22.
-//
-
-import UIKit
-
-extension UICollectionViewFlowLayout {
-    
-    var leftRightAndInteritem: CGFloat {
-        minimumInteritemSpacing + sectionInset.left + sectionInset.right
-    }
-}
