--- QRCodeTarot/Extensions/UICollectionViewLayout.swift
@@ -1,15 +1,0 @@
-//
-//  UICollectionViewLayout.swift
-//  QRCodeTarot
-//
-//  Created by Scott Lydon on 5/4/22.
-//
-
-import UIKit
-
-extension UICollectionViewLayout {
-
-    var asFlowLayout: UICollectionViewFlowLayout? {
-        self as? UICollectionViewFlowLayout
-    }
-}
