+++ QRCodeTarot/Extensions/UIColor.swift
