+++ QRCodeTarot/Extensions/UIEdgeInsets.swift
