+++ QRCodeTarot/Extensions/UIFont.swift
@@ -1,0 +1,15 @@
+//
+//  UIFont.swift
+//  QRCodeTarot
+//
+//  Created by Scott Lydon on 5/18/22.
+//
+
+import UIKit
+
+extension UIFont {
+
+    static func inter(size: CGFloat) -> UIFont {
+        UIFont(name: "Inter", size: size)!
+    }
+}
