+++ QRCodeTarot/Extensions/UILabel.swift
@@ -32,22 +32,4 @@ extension UILabel: HasViewModel {
             .init()
         }
     }
-
-    func with(text: String) -> Self {
-        self.text = text
-        return self
-    }
-
-    func with(lineCount: Int) -> Self {
-        self.numberOfLines = lineCount
-        return self
-    }
-
-    /// Call this no more than once per UILabel instance.
-    func setDynamicText() {
-        font = font.dynamic
-        adjustsFontForContentSizeCategory = true
-        setContentCompressionResistancePriority(.required, for: .horizontal)
-        setContentCompressionResistancePriority(.required, for: .vertical)
-    }
 }
