+++ QRCodeTarot/ViewControllers/CardDetailViewController/TarotSwitchView/TarotSwitchView.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 import TableMVVM
+import CommonUIKitExtensions
 
 class TarotSwitchView: NibView, HasViewModel {
