+++ QRCodeTarot/ViewControllers/SubMenuViewController/TutorialStepView/TutorialStepView.swift
@@ -45,9 +45,3 @@ extension TutorialStepView {
         }
     }
 }
-
-extension Int {
-    var string: String {
-        String(self)
-    }
-}
