+++ QRCodeTarot/ViewControllers/TarotCardsViewController/TarotCardsViewController.swift
@@ -34,30 +34,6 @@ class TarotCardsViewController: UIViewController {
     }
 }
 
-extension UIView {
-
-    func addGradient(toTop: NSNumber = 0.05) {
-        let gradient = CAGradientLayer()
-        assert(superview != nil, "make sure you add the view to the superview before adding the gradient.")
-        gradient.frame = superview?.bounds
-            .increasedWidth(by: 50) // idk why, but this prevents the chop off on the right side issue.
-        ?? CGRect.null
-        gradient.colors = [UIColor.clear.cgColor, UIColor.black.cgColor]
-        gradient.locations = [0.0, toTop]
-        superview?.layer.mask = gradient
-    }
-}
-
-extension CGRect {
-    func increasedWidth(by amount: CGFloat) -> CGRect {
-        .init(
-            x: self.minX,
-            y: self.minY,
-            width: self.width + amount,
-            height: self.height
-        )
-    }
-}
 
 extension TarotCardsViewController: UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
